video files moved to video folder
